[Refactor]: Moved non-state worrying components into presentational folder
