+++ taigit/src/components/presentational/TeamMemberCard.js
